IWS-RiskControl/requirements.txt
kafka-python==2.0.2
openai==1.82.0
